Extract QR mocking code to test-utils
