+++ apps/web/test/unit-tests/components/views/elements/QRCode-test.tsx
@@ -6,41 +6,15 @@ SPDX-License-Identifier: AGPL-3.0-only OR GPL-3.0-only OR LicenseRef-Element-Com
 Please see LICENSE files in the repository root for full details.
 */
 
-import { act, render, cleanup } from "jest-matrix-react";
+import { render, cleanup } from "jest-matrix-react";
 import React from "react";
-import { toDataURL, type QRCodeSegment, type QRCodeToDataURLOptions } from "qrcode";
 
+import { mockQRCodeRender, resetQRCodeMock, waitForQRCodeRender } from "../../../../test-utils/qrcode";
 import QRCode from "../../../../../src/components/views/elements/QRCode";
-
-jest.mock("qrcode", () => ({
-    ...jest.requireActual("qrcode"),
-    toDataURL: jest.fn(),
-}));
-
-const realQRCode = jest.requireActual("qrcode") as { toDataURL: typeof toDataURL };
-const mockedToDataURL = jest.mocked(toDataURL);
-
-let qrCodeRenderPromise: Promise<string>;
-
-function mockQRCodeRender(): void {
-    // Keep real PNG generation, but capture the promise so the test can await it directly.
-    mockedToDataURL.mockImplementation(((data: string | QRCodeSegment[], options?: QRCodeToDataURLOptions) => {
-        qrCodeRenderPromise = realQRCode.toDataURL(data, options);
-        return qrCodeRenderPromise;
-    }) as typeof toDataURL);
-}
-
-async function waitForQRCodeRender(): Promise<void> {
-    // Flush the React state update scheduled by QRCode after toDataURL resolves.
-    await act(async () => {
-        await qrCodeRenderPromise;
-        await Promise.resolve();
-    });
-}
 
 describe("<QRCode />", () => {
     afterEach(() => {
-        mockedToDataURL.mockReset();
+        resetQRCodeMock();
         cleanup();
     });
